streamlit_app/requirements.txt
streamlit
arxiv
langchain
langchain-community
langchain-huggingface
chromadb
tiktoken
sentence-transformers
pandas
google-cloud-storage 